+++ src/app/service/poap/StartPOAP.ts
@@ -39,7 +39,7 @@ export default async (ctx: CommandContext, guildMember: GuildMember, platform: s
 	
 	if (platform == constants.PLATFORM_TYPE_TWITTER) {
 		// await startTwitterPOAPFlow(ctx, guildMember, db, event);
-		await guildMember.send('coming soon');
+		await ctx.send('Coming soon...');
 		return;
 	}
 	
@@ -55,7 +55,7 @@ export default async (ctx: CommandContext, guildMember: GuildMember, platform: s
 		throw new ValidationError(`Please end the active event \`${activeSettings.voiceChannelName}\`.`);
 	}
 	
-	await ServiceUtils.tryDMUser(guildMember);
+	await ServiceUtils.tryDMUser(guildMember, 'Oh yea, time for a POAP event!...');
 	await guildMember.send({
 		content: 'For which voice channel should the POAP event occur?',
 	});